+++ vite/src/utils/index.ts
@@ -1,2 +1,3 @@
 // Export all utility functions
-export * from './layerzero-scan'; 
+export * from './layerzero-scan';
+export * from './network';
