+++ vite/src/utils/oft/index.ts
@@ -1,0 +1,2 @@
+// Solana OFT utilities
+export * from './solana';
